quote-api/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
pydantic
typing 
psycopg2_binary
python-dotenv
